+++ Cookbook/Database/Services/Client/ClientService.cs
@@ -2,9 +2,6 @@
 using System.Threading.Tasks;
 using Cookbook.Database.Repositories.Client;
 using Cookbook.Database.Services.Interfaces.ClientInterfaces;
-using Cookbook.Models;
-using Cookbook.Models.Database;
-using Cookbook.Models.Login;
 using Models.Models.Database;
 using ClientModel = Models.Models.Database.Client.Client;
 
@@ -17,16 +14,21 @@ public class ClientService : IClientService
     public ClientService()
     {
         _clientRepository = new ClientRepository();
-       }
-
+    }
     
     public async Task<ClientModel?> GetClientAsync(int id)
     {
+        if (id <= 0)
+            return null;
+        
         return await _clientRepository.GetClientAsync(id);
     }
 
     public async Task<ClientModel?> GetClientAsync(string login)
     {
+        if (string.IsNullOrEmpty(login))
+            return null;
+
         return await _clientRepository.GetClientAsync(login);
     }
 
@@ -37,16 +39,40 @@ public class ClientService : IClientService
 
     public async Task<CommandResult> AddClientAsync(ClientModel client)
     {
+        if(string.IsNullOrWhiteSpace(client.Login))
+            return CommandResults.BadRequest;
+        
+        if(string.IsNullOrWhiteSpace(client.Password))
+            return CommandResults.BadRequest;
+            
+        if(string.IsNullOrWhiteSpace(client.Name))
+            return CommandResults.BadRequest;
+
         return await _clientRepository.AddClientAsync(client);
     }
 
     public async Task<CommandResult> UpdateClientAsync(ClientModel client)
     {
+        if(client.Id <= 0)
+            return CommandResults.BadRequest;
+
+        if(string.IsNullOrWhiteSpace(client.Login))
+            return CommandResults.BadRequest;
+        
+        if(string.IsNullOrWhiteSpace(client.Password))
+            return CommandResults.BadRequest;
+            
+        if(string.IsNullOrWhiteSpace(client.Name))
+            return CommandResults.BadRequest;
+        
         return await _clientRepository.UpdateClientAsync(client);
     }
 
     public async Task<CommandResult> DeleteClientAsync(int id)
     {
+        if (id <= 0)
+            return CommandResults.BadRequest;
+
         return await _clientRepository.DeleteClientAsync(id);
     }
 }
